feat(deps): upgrade llm-kernel 0.1.0 → 0.1.1, delegate cosine f64 and OpenAI model dispatch
- Bump llm-kernel version pin to 0.1.1
- Remove redundant `as f64` cast in cosine_similarity: 0.1.1 already
  accumulates in f64, so the wrapper now returns the result directly
- Replace model.contains("large") heuristic in create_openai_provider
  with an explicit match; unknown model names now return a clear error
  instead of silently falling back to new_small

Closes #36

+++ src/adapters/infra/search_engines.rs
@@ -301,9 +301,9 @@ pub fn sparse_entity_boost(counts: &HashMap<String, usize>, entity_id: &str) -> 
 
 /// Compute cosine similarity between two f32 vectors.
 ///
-/// Delegates to [`llm_kernel::embedding::cosine_similarity`] (single-pass f32).
+/// Delegates to [`llm_kernel::embedding::cosine_similarity`] (f64-precision accumulation).
 pub fn cosine_similarity(a: &[f32], b: &[f32]) -> f64 {
-    llm_kernel::embedding::cosine_similarity(a, b) as f64
+    llm_kernel::embedding::cosine_similarity(a, b)
 }
 
 // ---------------------------------------------------------------------------
